requirements.txt
SpeechRecognition==3.10.0
pyaudio==0.2.14
pyttsx3==2.90
ollama==0.1.6
screen-brightness-control==0.16.1
pycaw==20181226
comtypes==1.2.0
psutil==5.9.8
python-dotenv
opencv-python
pyautogui
telegram
python-telegram-bot